StarpageComponent: Move it to the component folder
