+++ src/app/views/components/startpage/startpage.component.scss
