+++ src/app/views/components/startpage/startpage.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { environment } from '../../../environments/environment';
+import { environment } from '../../../../environments/environment';
 
 @Component({
     selector:    'tcd-startpage',
